+++ lib/odata/entity_set.rb
@@ -105,9 +105,15 @@ module OData
       url_chunk, options = setup_entity_post_request(entity)
       result = execute_entity_post_request(options, url_chunk)
       if entity.is_new?
-        doc = ::Nokogiri::XML(result).remove_namespaces!
-        entity[entity.primary_key] = doc.xpath("//content/properties/#{entity.primary_key}").first.content
+        doc = ::Nokogiri::XML(result.body).remove_namespaces!
+        primary_key_node = doc.xpath("//content/properties/#{entity.primary_key}").first
+        entity[entity.primary_key] = primary_key_node.content unless primary_key_node.nil?
       end
+
+      unless result.code.to_s =~ /^2[0-9][0-9]$/
+        entity.errors << ['could not commit entity']
+      end
+
       entity
     end
 
@@ -143,13 +149,27 @@ module OData
     def execute_entity_post_request(options, url_chunk)
       result = service.execute(url_chunk, options)
       unless result.code.to_s =~ /^2[0-9][0-9]$/
-        raise StandardError, 'Something went wrong committing your entity'
+        service.logger.debug <<-EOS
+          [ODATA: #{service_name}]
+          An error was encountered committing your entity:
+
+            POSTed URL:
+            #{url_chunk}
+
+            POSTed Entity:
+            #{options[:body]}
+
+            Result Body:
+            #{result.body}
+        EOS
+        service.logger.info "[ODATA: #{service_name}] Unable to commit data to #{url_chunk}"
       end
-      result.body
+      result
     end
 
     def setup_entity_post_request(entity)
-      chunk = entity.is_new? ? name : "#{name}(#{entity[entity.primary_key]})"
+      primary_key = entity.send(:properties)[entity.primary_key].url_value
+      chunk = entity.is_new? ? name : "#{name}(#{primary_key})"
       options = {
           method: :post,
           body: entity.to_xml.gsub(/\n\s+/, ''),